add build of Bash loadable with --enable-bash-loadable
src/main.c is also used to build the loadable. bash/ is somewhat served as build directory.

The test conditional in testsuite.td.at is changed to

    AT_SKIP_IF([test "x$3" != xyes])

Because installcheck would produce a "x" == xno check resulting a false and the test is not skipped, which is a problem for loadable since it's not build by default.

Also

    AT_CHECK([$2 -P -p 60]

Is changed to -p -- 60, that's because the different behaviors between GNU getopt and Bash's for options with optional arguments. Using -- to ensure the results are all the same.